// RESPONSIVE WEB DEVELOPMENT/src/interfaces/medico.interfaces.ts
export interface iUpdateMedico {
    nome: string;
    email: string;
    senha: string;
    cpf: string;
    crm: string;
    endereco: string;
    estado: string;
    cidade: string;
    cep: string;
}

export interface iUpdateDaysMedico {
    segunda: string;
    terca: string;
    quarta: string;
    quinta: string;
    sexta: string;
    sabado: string;
}